+++ src/util/config-reader.ts
@@ -107,49 +107,89 @@ export class ConfigReader {
      * @returns value of the item
      */
     get(key: string, defaultValue = null, loop?: string) {
+        if (!key) {
+            return null;
+        }
+        if (process && key in process.env) {
+            return process.env[key];
+        }
         const result = this.config.getElement(key, defaultValue);
-        if (typeof result == 'string') {
-            const bracketStart = result.indexOf('${');
-            const bracketEnd = bracketStart != -1? result.indexOf('}', bracketStart) : -1;
-            if (bracketStart != -1 && bracketEnd != -1 && bracketEnd > bracketStart) {                
-                let middle = result.substring(bracketStart + 2, bracketEnd).trim();
-                let middleDefault = null;
-                if (middle) {
-                    const loopId = String(loop? loop : util.getUuid());
-                    const colon = middle.lastIndexOf(':');
-                    if (colon > 0) {
-                        middleDefault = middle.substring(colon+1);
-                        middle = middle.substring(0, colon);
+        if (typeof result == 'string' && self) {
+            if (result.lastIndexOf('${') != -1) {
+                const segments = this.extractSegments(result);
+                segments.reverse();
+                let start = 0;
+                let sb = '';
+                for (const i in segments) {
+                    const s = segments[i];
+                    const middle = result.substring(s.start+2, s.end-1).trim();
+                    const evaluated = this.performEnvVarSubstitution(key, middle, defaultValue, loop);
+                    const heading = result.substring(start, s.start);
+                    if (heading) {
+                        sb += heading;
                     }
-                    if (process && middle in process.env) {
-                        middle = middle in process.env? process.env[middle] : null;
-                    } else {
-                        const refs: Array<string> = this.loopDetection.has(loopId)? this.loopDetection.get(loopId) as Array<string> : [];
-                        if (refs.includes(middle)) {
-                            log.warn(`Config loop for '${middle}' detected`);
-                            middle = "* config loop *";
-                        } else {
-                            refs.push(middle);
-                            this.loopDetection.set(loopId, refs);
-                            // "self" points to the base configuration
-                            middle = self.get(middle, defaultValue, loopId);
-                        }
+                    if (evaluated) {
+                        sb += evaluated;
                     }
-                    this.loopDetection.delete(loopId);
-                    const first = result.substring(0, bracketStart);
-                    const last = result.substring(bracketEnd+1);
-                    if (first || last) {
-                        if (middleDefault == null) {
-                            middleDefault = '';
-                        }
-                        return first + (middle == null? middleDefault : middle) + last;
-                    } else {
-                        return middle == null? middleDefault : middle;
-                    }                                    
+                    start = s.end;
                 }
+                const lastSegment = result.substring(start);
+                if (lastSegment) {
+                    sb += lastSegment;
+                }
+                return sb? sb : null;
             }
         }
         return result;
+    }
+
+    extractSegments(original: string): EnvVarSegment[] {
+        const result = [];
+        let text = original;
+        while (true) {
+            const bracketStart = text.lastIndexOf('${');
+            const bracketEnd = text.lastIndexOf('}');
+            if (bracketStart != -1 && bracketEnd != -1 && bracketEnd > bracketStart) {
+                result.push(new EnvVarSegment(bracketStart, bracketEnd+1));
+                text = original.substring(0, bracketStart);
+            } else if (bracketStart != -1) {
+                text = original.substring(0, bracketStart);
+            } else {
+                break;
+            }
+        }
+        return result;
+    }
+
+    performEnvVarSubstitution(key: string, text: string, defaultValue = null, loop?: string): string {
+        if (text) {
+            let middleDefault = null;
+            const loopId = String(loop? loop : util.getUuid());
+            const colon = text.lastIndexOf(':');
+            if (colon > 0) {
+                middleDefault = text.substring(colon+1);
+                text = text.substring(0, colon);
+            }
+            if (process && text in process.env) {
+                text = process.env[text];
+            } else {
+                const refs: Array<string> = this.loopDetection.has(loopId)? this.loopDetection.get(loopId) as Array<string> : [];
+                if (refs.includes(text)) {
+                    log.warn(`Config loop for '${key}' detected`);
+                    text = '';
+                } else {
+                    refs.push(text);
+                    this.loopDetection.set(loopId, refs);
+                    // "self" points to the base configuration
+                    const mid = self.get(text, defaultValue, loopId);
+                    text = mid? String(mid) : null;
+                }
+            }
+            this.loopDetection.delete(loopId);
+            return text? text : middleDefault;
+        } else {
+            return defaultValue? String(defaultValue) : null;
+        }
     }
 
     /**
@@ -190,5 +230,17 @@ export class ConfigReader {
         return this;
     }
 
+
+
+}
+
+class EnvVarSegment {
+    public start: number = 0;
+    public end: number = 0;
+
+    constructor(start: number, end: number) {
+        this.start = start;
+        this.end = end;
+    }
 }