+++ CHANGELOG.md
@@ -6,6 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.0.1, 12/16/2024
+
+### Added
+
+Support parsing of multiple environment variables and base system properties for
+a single key-value in Config Reader.
+
+### Removed
+
+N/A
+
+### Changed
+
+1. Improved environment variable parsing logic and detection of config loops.
+2. Compatibility with Unix, Mac and Windows OS
+
+---
 ## Version 4.0.0, 12/9/2024
 
 Upgraded to sync with Mercury-Composable for the foundation event-driven and Event-over-HTTP